requirements.txt
streamlit
pandas
openai
gspread
google-auth-oauthlib
rapidfuzz
requests
Pillow
numpy
python-docx
# --- LangChain & LangGraph Libraries ---
langchain
langchain-openai
langchain-community
faiss-cpu
tiktoken
langgraph

